Added GridGraphCreator which extracts water clusters.
Modified a lot of other code that didn't work or was bad.
QFrame had a bug where range was always 0.
FrameSplitter now doesn't need a QFrame to split up, as
it only requires information from the PDB file.
Thus WaterMolecule stores the indices of the OH2, H1 and H2
in the PDB file. At the moment the time taken to find all
clusters in all frames is about 4m15s for high accuracy(?)
and 2m15 for low accuracy.

The current metric is nearest neighbour in R3, which may not
be the best metric to use as we want to take into account the
angle as well.

See the main.cpp on how I use this crap.

The includes are still broken to be how they are in temporary
directory.