Interframe Nearest Neighbour Permutation Compressor encoding implemented.
This works by encoding permutation by relating it to the position of an atom
in the previous frame. The value that gets encoded per atom is to store the
distance rank of the position in the previous frame from the position in the
current frame.

Permutation compressor's now take an optional argument to next_index
specifying the position of the atom. Only watercomp has been modified to pass
this information.